remove .solo
