+++ test/js/samples/svelte-element-svg/expected.js
@@ -33,8 +33,8 @@ function create_dynamic_element(ctx) {
 
 	return {
 		c() {
-			svelte_element1 = svg_element("svg");
-			svelte_element0 = svg_element("path");
+			svelte_element1 = svg_element(/*tag*/ ctx[0].svg);
+			svelte_element0 = svg_element(/*tag*/ ctx[0].path);
 			set_svg_attributes(svelte_element0, svelte_element0_data);
 			set_svg_attributes(svelte_element1, svelte_element1_data);
 		},
@@ -55,9 +55,9 @@ function create_dynamic_element(ctx) {
 }
 
 function create_fragment(ctx) {
-	let previous_tag = "svg";
+	let previous_tag = /*tag*/ ctx[0].svg;
 	let svelte_element_anchor;
-	let svelte_element = "svg" && create_dynamic_element(ctx);
+	let svelte_element = /*tag*/ ctx[0].svg && create_dynamic_element(ctx);
 
 	return {
 		c() {
@@ -69,12 +69,12 @@ function create_fragment(ctx) {
 			insert(target, svelte_element_anchor, anchor);
 		},
 		p(ctx, [dirty]) {
-			if ("svg") {
+			if (/*tag*/ ctx[0].svg) {
 				if (!previous_tag) {
 					svelte_element = create_dynamic_element(ctx);
 					svelte_element.c();
 					svelte_element.m(svelte_element_anchor.parentNode, svelte_element_anchor);
-				} else if (safe_not_equal(previous_tag, "svg")) {
+				} else if (safe_not_equal(previous_tag, /*tag*/ ctx[0].svg)) {
 					svelte_element.d(1);
 					svelte_element = create_dynamic_element(ctx);
 					svelte_element.c();
@@ -87,7 +87,7 @@ function create_fragment(ctx) {
 				svelte_element = null;
 			}
 
-			previous_tag = "svg";
+			previous_tag = /*tag*/ ctx[0].svg;
 		},
 		i: noop,
 		o: noop,
@@ -98,10 +98,15 @@ function create_fragment(ctx) {
 	};
 }
 
+function instance($$self) {
+	const tag = { svg: 'svg', path: 'path' };
+	return [tag];
+}
+
 class Component extends SvelteComponent {
 	constructor(options) {
 		super();
-		init(this, options, null, create_fragment, safe_not_equal, {});
+		init(this, options, instance, create_fragment, safe_not_equal, {});
 	}
 }
